fix: rename createMMKV.mock.ts to createMockMMKV.ts to ensure the mock file doesn't conflict with custom platform extensions
